requirements.txt
click
elasticsearch
streamlit
plotly
unidecode
pycountry
cleanco
geonamescache
pandas
sentence-transformers
pyyaml
python-dotenv